Registro
Se ha añadido el registro en api y app

+++ Api_EcoClime/src/main/java/com/example/demo/CitaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CitaRepository extends JpaRepository<Cita, Long> {
-    List<Cita> findByUsuarioId(Long usuarioId); // Obtener todas las citas de un usuario
+public interface CitaRepository extends JpaRepository<Cita, Integer> {
+    List<Cita> findByUsuarioId(Integer usuarioId); // Obtener todas las citas de un usuario
 }
